prepare for AMD64 assembler implementations
This commit prepares the AES-GCM-SIV and AES-SIV-CMAC
implementation for specialized AMD64 assembly.

Therefore this commit moves the entire generic logic
for AES-* to *_noasm.go. It also adds the private
`authEnc` interface implemented by generic/platform-
dependent code.
