+++ go.mod
@@ -1,3 +1,6 @@
 module github.com/secure-io/siv-go
 
-require github.com/aead/cmac v0.0.0-20160719120800-7af84192f0b1
+require (
+	github.com/aead/cmac v0.0.0-20160719120800-7af84192f0b1
+	golang.org/x/sys v0.0.0-20180909124046-d0be0721c37e
+)
